Factor out SpeechSettingsFeature
