+++ count/Feature/SettingsFeature.swift
@@ -4,99 +4,41 @@ import SwiftUI
 
 struct SettingsFeature: Reducer {
     struct State: Equatable {
-        var availableVoices: [SpeechSynthesisVoice]
-        @BindingState var rawSpeechRate: Float
-        @BindingState var rawVoiceIdentifier: String?
-        @BindingState var rawPitchMultiplier: Float
-        let pitchMultiplierRange: ClosedRange<Float>
-        let speechRateRange: ClosedRange<Float>
-        var speechSettings: SpeechSynthesisSettings
+        var speechSettings: SpeechSettingsFeature.State
 
         let topic: Topic
 
-        init(topicID: UUID, speechSettings: SpeechSynthesisSettings) {
+        init(topicID: UUID, speechSettings: SpeechSettingsFeature.State = .init()) {
             @Dependency(\.speechSynthesisClient) var speechClient
             @Dependency(\.topicClient.allTopics) var allTopics
 
             topic = allTopics()[id: topicID]!
             self.speechSettings = speechSettings
-
-            availableVoices = speechClient.availableVoices()
-            rawVoiceIdentifier = speechSettings.voiceIdentifier ?? speechClient.defaultVoice()?.voiceIdentifier
-
-            let speechRateAttributes = speechClient.speechRateAttributes()
-            rawSpeechRate = speechSettings.rate ?? speechRateAttributes.defaultRate
-            speechRateRange = speechRateAttributes.minimumRate ... speechRateAttributes.maximumRate
-
-            let pitchMultiplierAttributes = speechClient.pitchMultiplierAttributes()
-            rawPitchMultiplier = speechSettings.pitchMultiplier ?? pitchMultiplierAttributes.defaultPitch
-            pitchMultiplierRange = pitchMultiplierAttributes.minimumPitch ... pitchMultiplierAttributes.maximumPitch
         }
     }
 
-    enum Action: BindableAction, Equatable {
-        case binding(BindingAction<State>)
+    enum Action: Equatable {
         case doneButtonTapped
-        case onSceneWillEnterForeground
         case onTask
-        case pitchLabelDoubleTapped
-        case rateLabelDoubleTapped
-        case testVoiceButtonTapped
+        case speechSettings(SpeechSettingsFeature.Action)
     }
 
     @Dependency(\.dismiss) var dismiss
-    @Dependency.Notification(\.sceneWillEnterForeground) var sceneWillEnterForeground
-    @Dependency(\.speechSynthesisClient) var speechClient
 
     var body: some ReducerOf<Self> {
-        BindingReducer()
+        Scope(state: \.speechSettings, action: /Action.speechSettings) {
+            SpeechSettingsFeature()
+        }
         Reduce { state, action in
             switch action {
-            case .binding(\.$rawSpeechRate):
-                state.speechSettings.rate = state.rawSpeechRate
-                return .none
-            case .binding(\.$rawVoiceIdentifier):
-                state.speechSettings.voiceIdentifier = state.rawVoiceIdentifier
-                return .none
-            case .binding(\.$rawPitchMultiplier):
-                state.speechSettings.pitchMultiplier = state.rawPitchMultiplier
-                return .none
-            case .binding:
-                return .none
             case .doneButtonTapped:
                 return .run { send in
                     await dismiss()
                 }
-            case .pitchLabelDoubleTapped:
-                state.rawPitchMultiplier = speechClient.pitchMultiplierAttributes().defaultPitch
-                state.speechSettings.pitchMultiplier = state.rawPitchMultiplier
-                return .none
-            case .rateLabelDoubleTapped:
-                state.rawSpeechRate = speechClient.speechRateAttributes().defaultRate
-                state.speechSettings.rate = state.rawSpeechRate
-                return .none
-            case .onSceneWillEnterForeground:
-                state.availableVoices = speechClient.availableVoices()
-                return .none
             case .onTask:
-                return .run { send in
-                    for await _ in sceneWillEnterForeground {
-                        await send(.onSceneWillEnterForeground)
-                    }
-                }
-            case .testVoiceButtonTapped:
-                let spokenText = "1234"
-                enum CancelID { case speakAction }
-                return .run { [settings = state.speechSettings] send in
-                    await withTaskCancellation(id: CancelID.speakAction, cancelInFlight: true) {
-                        do {
-                            let utterance = SpeechSynthesisUtterance(speechString: spokenText, settings: settings)
-                            try await speechClient.speak(utterance)
-                        } catch {
-                            assertionFailure(error.localizedDescription)
-                        }
-                    }
-                }
+                return .send(.speechSettings(.onTask))
+            case .speechSettings:
+                return .none
             }
         }
     }
@@ -126,71 +68,9 @@ struct SettingsView: View {
                             .font(.subheadline)
                     }
 
-                    Section {
-                        if let $unwrappedVoiceIdentifier = Binding(viewStore.$rawVoiceIdentifier) {
-                            Picker(selection: $unwrappedVoiceIdentifier) {
-                                ForEach(viewStore.availableVoices) { voice in
-                                    Text(voice.name)
-                                        .tag(Optional(voice.voiceIdentifier))
-                                }
-                            } label: {
-                                Text("Voice name")
-                            }
-                            .pickerStyle(.navigationLink)
-                            NavigationLink {
-                                GetMoreVoicesView()
-                            } label: {
-                                Text("Get more voices")
-                            }
-                            HStack {
-                                Text("Rate")
-                                    .onTapGesture(count: 2) {
-                                        viewStore.send(.rateLabelDoubleTapped)
-                                    }
-                                Slider(value: viewStore.$rawSpeechRate, in: viewStore.speechRateRange, step: 0.05) {
-                                    Text("Speech rate")
-                                } minimumValueLabel: {
-                                    Image(systemName: "tortoise")
-                                } maximumValueLabel: {
-                                    Image(systemName: "hare")
-                                }
-                            }
-                            HStack {
-                                Text("Pitch")
-                                    .onTapGesture(count: 2) {
-                                        viewStore.send(.pitchLabelDoubleTapped)
-                                    }
-                                Slider(value: viewStore.$rawPitchMultiplier, in: viewStore.pitchMultiplierRange, step: 0.05) {
-                                    Text("Pitch")
-                                } minimumValueLabel: {
-                                    Image(systemName: "dial.low")
-                                } maximumValueLabel: {
-                                    Image(systemName: "dial.high")
-                                }
-                            }
-                            Button {
-                                viewStore.send(.testVoiceButtonTapped)
-                            } label: {
-                                HStack(spacing: 10) {
-                                    Image(systemName: "person.wave.2")
-                                    Text("Test Voice")
-                                }
-                                .frame(maxWidth: .infinity, alignment: .center)
-                            }
-                        } else {
-                            NavigationLink {
-                                GetMoreVoicesView()
-                            } label: {
-                                HStack {
-                                    Text("Error: no voices found on device")
-                                        .foregroundStyle(Color.red)
-                                }
-                            }
-                        }
-                    } header: {
-                        Text("Voice Settings")
-                            .font(.subheadline)
-                    }
+                    SpeechSettingsSection(
+                        store: store.scope(state: \.speechSettings, action: { .speechSettings($0) })
+                    )
                 }
                 .navigationBarTitleDisplayMode(.inline)
                 .navigationTitle("Session")
@@ -218,7 +98,7 @@ struct SettingsView: View {
 
 #Preview {
     SettingsView(
-        store: Store(initialState: SettingsFeature.State(topicID: Topic.mockID, speechSettings: .mock)) {
+        store: Store(initialState: SettingsFeature.State(topicID: Topic.mockID)) {
             SettingsFeature()
                 ._printChanges()
         } withDependencies: { deps in
